backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
alembic==1.16.5
asyncpg==0.30.0
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
email-validator==2.2.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-json-logger==3.3.0
httpx==0.28.1
pytest==8.4.2
pytest-asyncio==1.2.0
